+++ task15/src/main/java/task15/nosql/domain/Genre.java
@@ -1,4 +1,4 @@
-package task10.nosql.domain;
+package task15.nosql.domain;
 
 import lombok.*;
