+++ task15/src/main/java/task15/nosql/repository/AuthorRepository.java
@@ -1,7 +1,7 @@
-package task10.nosql.repository;
+package task15.nosql.repository;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
-import task10.nosql.domain.Author;
+import task15.nosql.domain.Author;
 
 public interface AuthorRepository extends MongoRepository<Author, String> {
 }
